requirements.txt
openai
fastapi
uvicorn
gunicorn
pydantic
dotenv
python-dotenv
azure-identity
azure-keyvault-secrets
